+++ lib/plat/optee/lws-plat-optee.c
@@ -73,8 +73,8 @@ lws_now_usecs(void)
 }
 #endif
 
-int
-lws_get_random(struct lws_context *context, void *buf, int len)
+size_t
+lws_get_random(struct lws_context *context, void *buf, size_t len)
 {
 #if defined(LWS_WITH_NETWORK)
 	TEE_GenerateRandom(buf, len);
